Replaced all repeated queries with existing functions and moved the exisiting functions into two files 'upload' and download'
